feat(tiles): use reprojection/stitching to pull tiles
Use the reprojection/stitching features of geopackage-js to map a full, typical XYZ tilegrid to a
potentially problematic (for Cesium) custom tilegrid. This has the added benefit in both 2D and 3D
of showing the imagery on all zoom levels even outside the levels that are valid within the
GeoPackage due to resampling.